fixed derivate handling of PID
In cases where the input value stays constant, and the reference value
is changing, the previous implementation would lead to the derivative
value not being used.

To calculate the derivative, only the input value was used, instead of
the full error value (e.g. the reference - input).

Changed the implementation to store the error instead of the previous
input value, and use that to calculate the derivative.